feat: add error codes

+++ src/main/java/pl/polsl/league/club/ClubController.java
@@ -16,41 +16,46 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import pl.polsl.league.error.ResourceNotFoundException;
+
 @Controller
 @RequestMapping(path = "/club")
 public class ClubController {
-	
+
 	@Autowired
 	ClubRepository clubRepository;
-	
+
 	@PostMapping
-    public @ResponseBody ClubDTO addClub(@RequestBody Club club) {
-        Club savedClub = clubRepository.save(club);
-        return new ClubDTO(savedClub);
-    }
+	public @ResponseBody ClubDTO addClub(@RequestBody Club club) {
+		Club savedClub = clubRepository.save(club);
+		return new ClubDTO(savedClub);
+	}
 
-    @GetMapping
-    public @ResponseBody CollectionModel<ClubDTO> getAllClubs() {
-        List<ClubDTO> clubsDTO = StreamSupport.stream(clubRepository.findAll().spliterator(), false)
-                .map(ClubDTO::new)
-                .collect(Collectors.toList());
-        return CollectionModel.of(clubsDTO);
-    }
+	@GetMapping
+	public @ResponseBody CollectionModel<ClubDTO> getAllClubs() {
+		List<ClubDTO> clubsDTO = StreamSupport.stream(clubRepository.findAll().spliterator(), false).map(ClubDTO::new)
+				.collect(Collectors.toList());
+		return CollectionModel.of(clubsDTO);
+	}
 
-    @GetMapping("/{id}")
-    public @ResponseBody ClubDTO getClub(@PathVariable Integer id) {
-        Club club = clubRepository.findById(id).orElse(null);
-        return club != null ? new ClubDTO(club) : null;
-    }
+	@GetMapping("/{id}")
+	public @ResponseBody ClubDTO getClub(@PathVariable Integer id) {
+		Club club = clubRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Record not found with id: " + id));
+		return new ClubDTO(club);
+	}
 
 	@PutMapping
 	public @ResponseBody ClubDTO updateClub(@RequestBody Club club) {
 		Club updatedClub = clubRepository.save(club);
 		return new ClubDTO(updatedClub);
 	}
-	
+
 	@DeleteMapping("/{id}")
 	public @ResponseBody void deleteClub(@PathVariable Integer id) {
-	    clubRepository.deleteById(id);
+		if (!clubRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Record not found with id: " + id);
+		}
+		clubRepository.deleteById(id);
 	}
 }
