+++ src/main/java/pl/polsl/league/participant/ParticipantController.java
@@ -16,6 +16,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import pl.polsl.league.error.ResourceNotFoundException;
+
 @Controller
 @RequestMapping(path = "/participant")
 public class ParticipantController {
@@ -31,16 +33,17 @@ public class ParticipantController {
 
 	@GetMapping
 	public @ResponseBody CollectionModel<ParticipantDTO> getAllParticipants() {
-		List<ParticipantDTO> participantsDTO = StreamSupport.stream(participantRepository.findAll().spliterator(), false)
-				.map(ParticipantDTO::new)
+		List<ParticipantDTO> participantsDTO = StreamSupport
+				.stream(participantRepository.findAll().spliterator(), false).map(ParticipantDTO::new)
 				.collect(Collectors.toList());
 		return CollectionModel.of(participantsDTO);
 	}
 
 	@GetMapping("/{id}")
 	public @ResponseBody ParticipantDTO getParticipant(@PathVariable Integer id) {
-		Participant participant = participantRepository.findById(id).orElse(null);
-		return participant != null ? new ParticipantDTO(participant) : null;
+		Participant participant = participantRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Record not found with id: " + id));
+		return new ParticipantDTO(participant);
 	}
 
 	@PutMapping
@@ -51,6 +54,9 @@ public class ParticipantController {
 
 	@DeleteMapping("/{id}")
 	public @ResponseBody void deleteParticipant(@PathVariable Integer id) {
+		if (!participantRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Record not found with id: " + id);
+		}
 		participantRepository.deleteById(id);
 	}
 }
